Framework and single test case for notebooks
This is just a helper function to load and run a
ipynb file and return a list of the contents of any
errors that occur. The idea is to add a test case
for each notebook we want to test (a single notebook
is included). Functional tests are better off in
python (maybe with code copied from the examples).